// src/api/generated/models/listLlmCallsRequest.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ListLlmCallsRequest {
  /** @minimum 1 */
  pageNum?: number;
  /** @minimum 1 */
  pageSize?: number;
  startTime?: string;
  endTime?: string;
  scenario?: string;
  modelName?: string;
  status?: string;
  finishReason?: string;
  errorType?: string;
  appId?: string;
  taskId?: string;
}