// src/api/generated/models/llmCallLogInfo.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface LlmCallLogInfo {
  id?: string;
  scenario?: string;
  modelName?: string;
  responseId?: string;
  finishReason?: string;
  appId?: string;
  taskId?: string;
  status?: string;
  inputTokens?: string;
  outputTokens?: string;
  totalTokens?: string;
  durationMillis?: string;
  errorType?: string;
  errorMessage?: string;
  createdAt?: string;
}