// src/api/generated/models/llmMonitoringOverview.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface LlmMonitoringOverview {
  startTime?: string;
  endTime?: string;
  callTotal?: string;
  successRate?: number;
  averageDurationMillis?: string;
  inputTokens?: string;
  outputTokens?: string;
  totalTokens?: string;
  errorCount?: string;
}