// src/api/generated/models/llmTokenUsageTrend.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { LlmTokenUsageTrendSeries } from './llmTokenUsageTrendSeries';

export interface LlmTokenUsageTrend {
  startTime?: string;
  endTime?: string;
  stepSeconds?: string;
  series?: LlmTokenUsageTrendSeries[];
}